GUI frame and drawing cards
Added an image (cards.png) containing every card in an UNO deck. A GUI class creates a JFrame, the DeckImage class initializes this image and allows for retrieval of the image of any card, and the Coordinates class is used by DeckImage to locate the coordinates of a given card. I've also added a test for these functions. The coordinates class does not yet account for wild and wild draw four cards, and returns coordinates that are just a few pixels off, getting worse the further from the origin the card appears in the base image. I intend to fix these, as well as make DeckImage a singleton (as its image only needs to be initialized once.)