The GUI now updates appropriately when cards are played
There is still no turn system implemented, so it will always remain the turn of the first player. The deck still cannot be drawn from via the GUI, and I've found a new error in which I could not play any green cards on top of what appeared as a green +2 card.